refactor: move `IntervalProvider` trait to `provider.rs`
