+++ packages/desktop/src/providers/battery/provider.rs
@@ -13,7 +13,7 @@ use tokio::task::AbortHandle;
 
 use super::{BatteryProviderConfig, BatteryVariables};
 use crate::providers::{
-  interval_provider::IntervalProvider, variables::ProviderVariables,
+  provider::IntervalProvider, variables::ProviderVariables,
 };
 
 pub struct BatteryProvider {
