+++ packages/desktop/src/providers/host/provider.rs
@@ -6,7 +6,7 @@ use tokio::{sync::Mutex, task::AbortHandle};
 
 use super::{HostProviderConfig, HostVariables};
 use crate::providers::{
-  interval_provider::IntervalProvider, variables::ProviderVariables,
+  provider::IntervalProvider, variables::ProviderVariables,
 };
 
 pub struct HostProvider {
